+++ pkgs/source_span/test/highlight_test.dart
@@ -53,6 +53,14 @@ zip zap zop
   '"""));
   });
 
+  test("works for a point span at the beginning of the file", () {
+    expect(file.location(0).pointSpan().highlight(), equals("""
+  ,
+1 | foo bar baz
+  | ^
+  '"""));
+  });
+
   test("works for a point span at the end of a line", () {
     expect(file.location(11).pointSpan().highlight(), equals("""
   ,
@@ -69,7 +77,26 @@ zip zap zop
   '"""));
   });
 
+  test("works for a point span after the end of the file", () {
+    expect(file.location(39).pointSpan().highlight(), equals("""
+  ,
+3 | zip zap zop
+  |            ^
+  '"""));
+  });
+
   test("works for a point span at the end of the file with no trailing newline",
+      () {
+    file = new SourceFile.fromString("zip zap zop");
+    expect(file.location(10).pointSpan().highlight(), equals("""
+  ,
+1 | zip zap zop
+  |           ^
+  '"""));
+  });
+
+  test(
+      "works for a point span after the end of the file with no trailing newline",
       () {
     file = new SourceFile.fromString("zip zap zop");
     expect(file.location(11).pointSpan().highlight(), equals("""
@@ -157,7 +184,7 @@ zip zap zop
   '"""));
     });
 
-    test("highlights the full last line if it's empty", () {
+    test("highlights the full first line if it's empty", () {
       var file = new SourceFile.fromString("""
 foo